These changes are relatively minor but do give some performance and memory improvements.

Summary of Changes
* Switch from deprecated `ANTLRInputStream` to antlr `CharStreams` approach
* Add new Parser functions that eliminate unnecessary copying of input
* Remove unnecessary traversals when transforming ANTLR CST to FIRRTL AST in the Visitor

For a ~700 MB input .fir file, this improves ANTLR parsing time from ~78s to ~52s and Visiting time from ~62s to ~54s. Modest but real improvements. The size of the ANTLR CST remains around ~36GB so while this helps reduce unnecessary object creation, the fundamental motivation for https://github.com/freechipsproject/chisel3/pull/829 remains